+++ internal/discovery/discoverer.go
@@ -13,11 +13,14 @@ type Replacement struct {
 	ReplaceField string `json:"replaceField"`
 }
 
+type VariablesMap map[string]interface{}
+type AnnotationsMap = VariablesMap
+
 // DiscoveredItem defines the structure of a single item that has been "discovered"
 type DiscoveredItem struct {
-	Variables         map[string]string `json:"variables"`
-	MetricAnnotations map[string]string `json:"metricAnnotations"`
-	EntityRewrites    []Replacement     `json:"entityRewrites"`
+	Variables         VariablesMap   `json:"variables"`
+	MetricAnnotations AnnotationsMap `json:"metricAnnotations"`
+	EntityRewrites    []Replacement  `json:"entityRewrites"`
 }
 
 // Output defines the final output of the discovery executable
@@ -48,7 +51,7 @@ func processContainers(containers []kubernetes.ContainerInfo) Output {
 	output := Output{}
 	for _, c := range containers {
 		// new map for each container
-		var discoveredProperties = make(map[string]string)
+		var discoveredProperties = make(VariablesMap)
 
 		discoveredProperties[namespace] = c.Namespace
 		discoveredProperties[podName] = c.PodName
@@ -63,6 +66,7 @@ func processContainers(containers []kubernetes.ContainerInfo) Output {
 		discoveredProperties[id] = c.ID
 		discoveredProperties[name] = c.Name
 		discoveredProperties[image] = c.Image
+		discoveredProperties[ports] = c.Ports
 		// although annotation are set in the pods, we "apply" them to containers
 		for k, v := range c.PodAnnotations {
 			discoveredProperties[annotationPrefix+k] = v
@@ -92,11 +96,11 @@ func getReplacements() []Replacement {
 }
 
 var annotationExclusions = []string{
-	id, ip, nodeIP,
+	id, ip, nodeIP, ports,
 }
 
-func filterAnnotations(props map[string]string) map[string]string {
-	filtered := make(map[string]string)
+func filterAnnotations(props VariablesMap) AnnotationsMap {
+	filtered := make(map[string]interface{})
 	for k, v := range props {
 		if strings.HasPrefix(k, annotationPrefix) {
 			continue